working solution

+++ lib/tx_history_buffer.h
@@ -19,7 +19,7 @@ template <typename Complex, std::size_t HistorySize> class TxHistoryBuffer {
     }
 
     void clear() noexcept {
-        samples_.fill(Complex(0.0f, 0.0f));
+        samples_.fill(Complex(0.0, 0.0));
     }
 
     void store(size_t sample_index, Complex sample) noexcept {
@@ -31,29 +31,49 @@ template <typename Complex, std::size_t HistorySize> class TxHistoryBuffer {
             store(start_index + i, samples[i]);
         }
     }
-
     Complex delayedSample(size_t sample_index, double delay_samples) const noexcept {
-        const float delayed_index =
-            static_cast<float>(sample_index) - static_cast<float>(delay_samples);
-        if (delayed_index < 0.0f) {
-            return Complex(0.0f, 0.0f);
+        const double delayed_index = static_cast<double>(sample_index) - delay_samples;
+        if (delayed_index < 0.0) {
+            return Complex(0.0, 0.0);
         }
 
         const size_t lower_index = static_cast<size_t>(std::floor(delayed_index));
         const size_t upper_index = lower_index + 1U;
         if (sample_index < lower_index || sample_index - lower_index >= HistorySize) {
-            return Complex(0.0f, 0.0f);
+            return Complex(0.0, 0.0);
         }
 
-        const float fraction = delayed_index - static_cast<float>(lower_index);
+        const double fraction = delayed_index - static_cast<double>(lower_index);
+
         const Complex lower_sample = samples_[lower_index & kMask];
         Complex upper_sample = lower_sample;
         if (upper_index <= sample_index && sample_index - upper_index < HistorySize) {
             upper_sample = samples_[upper_index & kMask];
         }
 
-        return ((1.0f - fraction) * lower_sample) + (fraction * upper_sample);
+        return ((1.0 - fraction) * lower_sample) + (fraction * upper_sample);
     }
+    // Complex delayedSample(size_t sample_index, double delay_samples) const noexcept {
+    //     const double delayed_index = static_cast<double>(sample_index) - delay_samples;
+    //     if (delayed_index < 0.0) {
+    //         return Complex(0.0, 0.0);
+    //     }
+    //
+    //     const size_t lower_index = static_cast<size_t>(std::floor(delayed_index));
+    //     const size_t upper_index = lower_index + 1U;
+    //     if (sample_index < lower_index || sample_index - lower_index >= HistorySize) {
+    //         return Complex(0.0f, 0.0f);
+    //     }
+    //
+    //     const double fraction = delayed_index - static_cast<double>(lower_index);
+    //     const Complex lower_sample = samples_[lower_index & kMask];
+    //     Complex upper_sample = lower_sample;
+    //     if (upper_index <= sample_index && sample_index - upper_index < HistorySize) {
+    //         upper_sample = samples_[upper_index & kMask];
+    //     }
+    //
+    //     return ((1.0 - fraction) * lower_sample) + (fraction * upper_sample);
+    // }
 
   private:
     std::array<Complex, HistorySize> samples_{};
